requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiohttp>=3.9.5
asyncpg>=0.29.0
redis[asyncio]>=5.0.4
pydantic>=2.7.1
python-dotenv>=1.0.1

# CLI / local mode
aiosqlite>=0.20.0
pyyaml>=6.0
rich>=13.7.0
prompt_toolkit>=3.0.46
gitpython>=3.1.40

# Optional: PNG export
matplotlib>=3.8.0; extra == "full"
pillow>=10.0.0; extra == "full"
numpy>=1.26.0; extra == "full"

# Optional: Telegram bot
python-telegram-bot[all]>=21.0; extra == "bot"
aiohttp-socks>=0.8.0; extra == "bot"
